requirements.txt
django
django-prose